+++ RELEASE_NOTES.md
@@ -1,3 +1,5 @@
+# Next Release
+
 # Release v0.11.0
 
 ## Highlights